update sortpart
added cluster diameters to get-num-clusts, but most changes made to clust-sig:
It now calculates statistics only up to the last peak which yields an M-value representing
at least one half of the total new paritions. The results plotted in fig3 of the
urban-clusters paper provide immediate support that this is an appropriate modification:
previous relationships between Delta G, N, and M were a bit all over the shop, yet with
this modification, they are pretty much all significantly and strongly positive.